route fixes, channel modification and deletion
