+++ pages/api/v1/[serverId]/channels/create.ts
@@ -1,21 +1,21 @@
-import { createChannel } from '@/services/channels.service';
+import { createChannel, updateChannel } from '@/services/channels.service';
 import type { NextApiRequest, NextApiResponse } from 'next';
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
   const { method } = req;
+  // TODO: Require user to have permission to manage channels
+  let serverId: number;
 
+  try {
+    serverId = parseInt(req.query.serverId as string);
+  }
+
+  catch (err: any) {
+    return res.status(400).send({ error: 'Invalid server ID' });
+  }
+
+  // NOTE: All operations after this point require appropriate authorization
   if (method === 'POST') {
-    // TODO: Require user to have permission to create channels
-    let serverId: number;
-
-    try {
-      serverId = parseInt(req.query.serverId as string);
-    }
-
-    catch (err: any) {
-      return res.status(400).send({ error: 'Invalid server ID' });
-    }
-
     await createChannel(
       serverId,
       req.body.name,
